rename internal/kit/... to internal/platform/...
