+++ internal/platform/msg/msg.go
